file_watcher/requirements.txt
google-cloud-storage==2.18.0
pytest==8.3.2
uuid==1.30
google-cloud-logging==3.11.0
pytz==2024.1
